requirements.txt
django==1.11
psycopg2
business-calendar==0.2.1
django-flat-responsive
django-allauth
pyyaml
cryptography